update tutorial steps 8-9 & demo 3
