--- demos/03-track-changes/appsettings.json.example
@@ -1,14 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Warning"
-    }
-  },
-  "MyConfig":
-  {
-    "AppId": "<APP ID>",
-    "AppSecret": "<APP SECRET>",
-    "TenantId": "<TENANT ID>",
-    "Ngrok": "<NGROK URL>"
-  }
-}
