+++ tutorial/09_completed.md
@@ -1,6 +1,6 @@
 <!-- markdownlint-disable MD002 MD041 -->
 
-You've completed the Microsoft Graph – Change Notifications and Track Changes tutorial. Now that you have a working app that calls Microsoft Graph, you can experiment and add new features. Visit the [Overview of Microsoft Graph](/graph/overview) to see all of the data you can access with Microsoft Graph.
+You've completed the Microsoft Graph – Change Notifications and Track Changes tutorial. Now that you have a working app that calls Microsoft Graph, you can experiment and add new features. Visit the [Overview of Microsoft Graph](https://docs.microsoft.com/graph/overview) to see all of the data you can access with Microsoft Graph.
 
 ## Feedback
